Renamed SwaggerUI to RedocUI for clarity, as the UI uses Redoc and not SwaggerUI
